use new remove duplicates function in ascwds not cqc
