+++ jobs/clean_ascwds_workplace_data.py
@@ -3,6 +3,7 @@ import sys
 from pyspark.sql import DataFrame
 from pyspark.sql.types import IntegerType
 
+from pyspark.sql.window import Window
 import pyspark.sql.functions as F
 
 from utils import utils
@@ -42,6 +43,8 @@ def main(source: str, destination: str):
         ascwds_workplace_df, [AWP.total_staff, AWP.worker_records]
     )
 
+    ascwds_workplace_df = remove_locations_with_duplicates(ascwds_workplace_df)
+
     ascwds_workplace_df = add_purge_outdated_workplaces_column(
         ascwds_workplace_df, AWPClean.ascwds_workplace_import_date
     )
@@ -64,6 +67,20 @@ def cast_to_int(df: DataFrame, column_names: list) -> DataFrame:
     for column in column_names:
         df = df.withColumn(column, df[column].cast(IntegerType()))
     return df
+
+
+def remove_locations_with_duplicates(df: DataFrame):
+    loc_id_import_date_window = Window.partitionBy(
+        AWP.location_id, PartitionKeys.import_date
+    )
+
+    df_with_count = df.withColumn(
+        "location_id_count", F.count(AWP.location_id).over(loc_id_import_date_window)
+    )
+
+    df_without_duplicates = df_with_count.filter(F.col("location_id_count") == 1)
+
+    return df_without_duplicates.drop("location_id_count")
 
 
 def add_purge_outdated_workplaces_column(
