+++ tests/unit/test_clean_cqc_location_data.py
@@ -261,48 +261,5 @@ class SplitDataframeIntoRegAndDeRegTests(CleanCQCLocationDatasetTests):
             self.assertEqual(warnings_log, [])
 
 
-class RemoveLocationsWithDuplicatesTests(CleanCQCLocationDatasetTests):
-    def setUp(self) -> None:
-        super().setUp()
-        self.returned_df = job.remove_locations_with_duplicates(self.test_location_df)
-
-        self.test_duplicate_loc_df = self.spark.createDataFrame(
-            Data.location_rows_with_duplicates, Schemas.small_location_schema
-        )
-
-    def test_returns_a_dataframe(self):
-        self.assertEqual(type(self.returned_df), DataFrame)
-
-    def test_returns_the_correct_columns(self):
-        self.assertCountEqual(
-            self.returned_df.columns, ["locationId", "providerId", "import_date"]
-        )
-
-    def test_does_not_remove_rows_if_no_duplicates(self):
-        self.assertEqual(self.returned_df.count(), self.test_location_df.count())
-        self.assertEqual(self.returned_df.collect(), self.test_location_df.collect())
-
-    def test_removes_duplicate_location_id_with_same_import_date(self):
-        filtered_df = job.remove_locations_with_duplicates(self.test_duplicate_loc_df)
-        expected_df = self.spark.createDataFrame(
-            Data.expected_filtered_location_rows, Schemas.small_location_schema
-        )
-        self.assertEqual(filtered_df.collect(), expected_df.collect())
-
-    def test_does_not_remove_duplicate_location_id_with_different_import_dates(self):
-        locations_with_different_import_dates_df = self.spark.createDataFrame(
-            Data.location_rows_with_different_import_dates,
-            Schemas.small_location_schema,
-        ).orderBy(CQCL.location_id)
-
-        filtered_df = job.remove_locations_with_duplicates(
-            locations_with_different_import_dates_df
-        ).orderBy(CQCL.location_id)
-
-        self.assertEqual(
-            filtered_df.collect(), locations_with_different_import_dates_df.collect()
-        )
-
-
 if __name__ == "__main__":
     unittest.main(warnings="ignore")
